+++ lib/simple_ads_manager.dart
@@ -1,3 +1,4 @@
 library simple_ads_manager;
 
 export 'src/ads_manager.dart';
+export 'src/models/ad_type.dart';